Move logging into own module
