+++ lib/ain-grpc/src/lib.rs
@@ -7,6 +7,7 @@ pub mod call_request;
 pub mod codegen;
 mod filters;
 mod impls;
+pub mod logging;
 mod receipt;
 pub mod rpc;
 mod sync;
@@ -20,6 +21,7 @@ use jsonrpsee::http_server::HttpServerBuilder;
 
 #[allow(unused)]
 use log::{debug, info};
+use logging::CppLogTarget;
 
 use crate::rpc::{
     debug::{MetachainDebugRPCModule, MetachainDebugRPCServer},
@@ -64,33 +66,12 @@ pub fn add_grpc_server(_runtime: &Runtime, _addr: &str) -> Result<(), Box<dyn Er
     Ok(())
 }
 
-struct LogTarget {}
-
-// TODO:
-// - Switch to u8 to avoid intermediate string conversions
-// - Use a custom format to clear up timing.
-impl std::io::Write for LogTarget {
-    fn write(&mut self, buf: &[u8]) -> std::io::Result<usize> {
-        // let mut out = std::io::stdout().lock();
-        // out.write(buf)?;
-        let s = std::str::from_utf8(buf).or(Err(std::io::Error::new(
-            std::io::ErrorKind::Other,
-            "Invalid UTF-8 sequence",
-        )))?;
-        ain_cpp_imports::log_print(s);
-        Ok(buf.len())
-    }
-
-    fn flush(&mut self) -> std::io::Result<()> {
-        Ok(())
-    }
-}
-
 pub fn preinit() {
     env_logger::Builder::from_env(
         env_logger::Env::default().default_filter_or(log::Level::Info.as_str()),
     )
-    .target(env_logger::Target::Pipe(Box::new(LogTarget {})))
+    .format(logging::cpp_log_target_format)
+    .target(env_logger::Target::Pipe(Box::new(CppLogTarget::new(false))))
     .init();
     info!("init");
 }
